requirements.txt
requests
python-dotenv
reportlab
signxml
cryptography
selenium
